+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+ * Fix parsing space-less media query features like `@media(width:123px)` [#141](https://github.com/premailer/css_parser/pull/141)
+
 ### Version v1.15.0
 
  * Fix parsing background shorthands in ruby 3.2 [#140](https://github.com/premailer/css_parser/pull/140)